Fix the check for -XFooBar ghc-options flags to be more permissive
Previously we rejected all such flags but that posed the problem that older
versions of Cabal, like 1.1.6 did not understand new extensions so we
could not actually follow the advice and use the extenion. So now we only
warn about -X flags if they refer to old extensions that Cabal 1.1.6 knew
about. If the .cabal file specifies cabal-version: >= 1.2 or similar
(anything that excludes 1.1.6) then we warn about all -X flags.

20080423190243-adfee-d3ff721f57b1b03aa7deee79e3a1e98f87c9fadb.gz