Cope better with ghc bug #2201, display a better error message
Otherwise it can (and does) really confuse people.
The problem is that the command $ ghc-pkg-6.9 describe '*' --user
returns a non-zero exit code if the user package db is empty.
ghc-pkg intends this exit code to tell us if the query returned
any results (one can use more complex queries as tests) but Cabal
interprets it as failure. Indeed we cannot distinguish it from
any other kind of failure from ghc-pkg.

20080505085746-adfee-30b21f3f434bdf43f0dea4fe3d1baf5c8f33006d.gz