Rename doc/fptools.css to avoid the ghc build system cleaning it
The user guide gets built in two different ways. There's a target
in Cabal's the top level Makefile and there is also the stuff that
the ghc build system uses. The ghc build system expects to copy in
doc/fptools.css and then delete it again on make clean. We want a
persistent copy so that we can make the docs when we've just got a
standalone Cabal build tree, so that's now kept as doc/Cabal.css.

20080520191700-adfee-4f3c670a157e404dc3d7a6de204e52d802a9aeb0.gz