Lift the restriction that libraries must have exposed-modules
This allows libs that have only private modules or C code. This
might be used to make libs that have non-exposed modules and only
export C APIs. It could also be used to make packages that consist
only of C code. That might be useful for bindings where it may
make sense to split the C and Haskell code into separate packages.

20080612092924-adfee-a737428f6a8c1fed15842d9ebf3cf4cedb7fbf6a.gz