'.' should not always be in hs-source dirs
We changed the parsing of list fields in the .cabal file so that it
adds to the current value rather than replacing it. This allows you
to put multiple entries for a list field and they all get
concatenated. However that means that the '.' in the hsSourceDirs of
emptyBuildInfo is always added to and not replaced like we did
previously. That's not what we want in this case. We want to use '.'
for hsSourceDirs *only* if hsSourceDirs is otherwise null. As it
happens, due to the way the configurations code works, we're already
filling in the default if it'd otherwise be null so we do not need
'.' in the emptyBuildInfo at all.

20080613230352-adfee-e043d46c87178ded93a5962d497e7dbf89c05fac.gz