Add PackageSet, like PackageIndex but case sensitive
Actually it turns out that we don't need case insensitivity in many
cases, mosty just for simple lookups in the UI. For everything else
the ordinary Ord instance is much simpler. The fact that listing the
contents of a PackageIndex doesn't come out in Ord order actually
causes real problems in cabal-install and necessitates re-sorting.
So we should move to using PackageSet in most cases and just leave
the search and lookup operations in PackageIndex.

20080614003705-adfee-12cead0af66f2d1d05e7488bbee103251f58b83f.gz