Always use gcc as cc with hsc2hs
Lookup what flags to use from the package index. Previously this
was done by calling ghc as cc and passing -package flags to ghc.
ghc would then lookup what extra flags to pass to gcc. We now do
that ourselves directly and it's a good deal simpler and it's
portable to the other haskell implementations. This is only a
first go, the flags may not all be exactly right. Needs testing.

20080720233759-adfee-afc666f51c90c7f2922b4a85778f5905dbd5b532.gz