Fix the semantics of the simple file globbing to be sane
I realised when I started to document it that the behaviour
was not terribly consistent or sensible. The meaning now is:
  The limitation is that * wildcards are only allowed in
  place of the file name, not in the directory name or
  file extension. In particular, wildcards do not include
  directories contents recursively. Furthermore, if a
  wildcard is used it must be used with an extension, so
  "data-files: data/*" is not allowed. When matching a
  wildcard plus extension, a file's full extension must
  match exactly, so "*.gz" matches "foo.gz" but not
  "foo.tar.gz".
The reason for providing only a very limited form of wildcard
is to concisely express the common case of a large number of
related files of the same file type without making it too easy
to accidentally include unwanted files.

20080729152624-adfee-82b21496ea160565e8dbb61f15c7fe50bdc7b656.gz