Add an "exposed" field to the .cabal file library section
It's a bool flag that says if by default the library should
be registered with the compiler as exposed/unhidden (for
compilers which have such a concept, ie ghc). You might want
to do this for packages which would otherwise pollute the
module namespace or clash with other common packages.
It should be very rarely used. The only current examples we
know of are the ghc api package and the dph packages.

20080730164516-adfee-3a2bc57dbe07445ab81cc5b32042ada7a09f5fc6.gz