Don't propagate cc-options to the InstalledPackageInfo
cc-options is for options to be passed to C compilations in the
current package.  If we propagate those options to the
InstalledPackageInfo, they get passed to C compilations in any package
that depends on this one, which could be disastrous.  I've seen
cc-options like these:

   cc-options:      -optc-std=c99
   cc-options:         -D_FILE_OFFSET_BITS=64
   Cc-options:      -Wall

these are all clearly intended to be local, but are in fact currently
propagated to all dependent packages.

20080821132551-12142-e4ddc118b2c842da0f977f73a13c627fd0ed630a.gz