Add support for manually en/disabled flags
The immediate use for these is so that, in haddock, we can require
ghc-paths normally, but in the GHC build we can manually turn off a flag
so that this dependency isn't needed. We can't use a normal flag, or
in the normal build Cabal would infer that the flag needs to be turned
off if ghc-paths isn't available.

20080827170105-3fd76-1c25affd29509e13ce196e164a9bf4d60cbea3c1.gz