Simplify the handling of --with-prog= in build/haddock commands.
We allow extra rgs and the location of programs to be given to
the build and haddock commands, not just at configure time. The
code to do this is now simpler and more general. This should not
be the default use mode however since it involves configuring
the programs each time where as doing it at configure time allows
it to be done once and saved. Further, specifying a different
version of the program at build time than at configure time is
likely to fail, especially for the compiler programs. Changing
the compiler really requires reconfiguring.

20080831215551-adfee-3f0f25569a28ae7987f29bd9aea6a440c2d23991.gz