Yet another go at making gcc -B work properly on windows
This time it should work on linux too! But more significantly
it should work when the user specifies a particular gcc. It
would be very bad if the user gave an alternative gcc but we
still gave it -B for the lib files of ghc's gcc. This go is
rather cleaner as it uses the new program post-conf system.

20080916232553-adfee-0fa82dbf30acc156b4d2a9458b4c81133b04d725.gz